최근후기 프래그먼트 작업  - 어댑터 + 데이터베이스 연결 / 어댑터 새로고침 순서(위치)확인하기  - 실제 데이터 반영

+++ app/src/main/java/com/start/mindcafeclonepractice/datas/ReviewData.kt
@@ -7,6 +7,6 @@ data class ReviewData(
     val solutionImg: String? = null,
     val content: String? = null,
     val ratingImg: String? = null,
-    val ratingNum: Float? = null
+    val ratingNum: Int? = null
 ) {
 }
